[driver_runtime] Simplify thread start conditions on dispatchers
Refactors the thread start logic to be less spread out through
the entire dispatcher object. Specifically:

- Remove the thread adder from `CreateWithAdder` and moves it
to be private, renames it, and moves the "should I spawn
a thread?" question to the `ThreadPool::OnDispatcherAdded`.
- Use the same logic for calculating max threads in both
dynamic and non-dynamic thread spawning.
- Renames `RemoveThread` to `OnDispatcherSealed` since that's
the context it's called in and it doesn't actually remove
a thread (yet).
- Renames `max_threads_` to `thread_limit_` (and related
functions) to reflect that the max is more complicated and
the naming of this configurable value in the public API.
- Makes `AddThread` just always add a thread, and adds
a locked version of the function so things don't have
to unlock before starting a new thread when they're
already locked. This avoids the need for a boolean
parameter that was not really correct (since it made
AddThread responsible for maintaining the count of
sync call dispatchers).
- Undoes various places where asserts for the number of
sync call dispatchers were checked for being >0 because
the accounting is now done more correctly.

This also unifies the logic for the default and non-default
thread pools and resolves b/42076454 in the process.

This reduces the number of threads on a minimal build after
boot from 129 to 113.

Fixed: 42076454
Change-Id: Ie6c5106a023d61d3fc074c5f54230c7d884fecfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1436414
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>