[fxfs] Fix race while force-locking in flush
If a Crypt error occurs while flushing a volume, we still want flushing
to proceed, so we force-lock the volume and retry.

There was a race condition which arose if this happened at the same time
that the volume was being unmounted, which could cause force-lock to
fail.

Fix this by making force-lock tolerant of the volume already being
locked.

Fixed: 457763372
Multiply: fxfs-platform-test.cm
Change-Id: If1069d853b9365d9857e1008df391fe4cf1987ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441513
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>