[cml] rewrite spanned -> context
yet another redo of my still not living code.
the inspiration was because the current merging
algorithm raises validation errors which my Spanned*
types cannot effectively catch.

The Spanned* types are replaced by Parsed* and Context*.
The Parsed* of each type is an intermediate step added
for parsing json. The Context* naming is the final form
that validation interacts with and uses to raise errors.

Change-Id: Ied009b143590414421d5b3f03e6740dd3235192d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440408
Commit-Queue: Olivia Nordquist <nolivia@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>