[input] InteractionStateHandler consults baton passing flags
The feature flags impact whether or not we eagerly retrieve a wake lease
and store it in a timer, based on the type of the input event (button,
mouse, or touch). These flags are currently all disabled (set to false)
on ToT, so there is no immediate behavioral change introduced.

Fixed: 465810142
Test: manual test shows behavior is consistent (suspend/resume works)
Test: manual test with flags locally enabled show the debug logs emitted
(given that upstream drivers do not yet send any wake leases)

Change-Id: Ifdba084feb2066d2cbcfaab35cdf36523e4e906e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440330
Commit-Queue: Caroline Liu <carolineliu@google.com>
Reviewed-by: Jianpeng Chao <chaopeng@google.com>