[recovery] Migrate from PerformReboot to Shutdown
This is a behavioral no-op.

PerformReboot, RebootToBootloader, RebootToRecovery, and Poweroff are
deprecated in favor of Shutdown. Compared to the deprecated methods,
Shutdown allows ALL clients to specify the reason for the shutdown.
Previously, only PerformReboot allowed clients to specify the shutdown
reason.

Feedback will use the additional information to improve the
determination of its canonical reboot reason, including but not limited
to filing crash reports for high temperature poweroffs.

Bug: 465530639
Test: `fx test //src/recovery`
Test: Booted into recovery, performed FDR. Reboot observed.
Change-Id: Ie7f210b2e2ad1126471a816fd4261b4162db556b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1442598
Commit-Queue: Thomas Gales <tgales@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>