[assembly][vbmeta] Support rollback indices
This change adds support for constructing a VBMeta with a specific
rollback index. We do not need this for Fuchsia VBMeta (and have adopted
no release convention around it), but we need to have a non-zero index
for VBMeta that we construct for the Android pVM context, even when
non-rollback-protected contexts. When we eventually add assembly logic
for such VBMeta, we'll just simply specify a 1 for now.

Bug: 449201661
Change-Id: Ic7c43a5f3b6de717bfed7faa14c1f32633a101ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425320
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>