[fx] use git-polyfill when needed.
This change will look to see if we are not running in a standard git
environment and if we aren't it will add the git-polyfill to the PATH.
The change makes it possible for tools like `fx build` to use the git
polyfill without having to change the build files. Additionally, it will
let us use git from inside other fx commands like `fx format-code`
without needing to change them.

Bug: 461899545
Change-Id: Idb3bbf3a083ad23a094a03f0ae4cbb767f11d390
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1439164
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>