[ffx][target] Add target package gc
This adds ffx support to trigger a garbage collection on a target. It
is intended to be a drop-in replacement for running the command `fx
shell pkgctl gc`.
For use by ffx, the previous internal fidl fuchsia.space is renamed to
fuchsia.pkg.garbagecollector and offered to toolbox.

Bug: 465722149
Change-Id: I971d1ac4983a82918482c28d9f8e50b06f1642f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1436016
Reviewed-by: Ben Keller <galbanum@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>
API-Review: Ben Keller <galbanum@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>