[netlink] Retry RuleSet creation
Netlink installs the rule set at a predefined priority. However, the
cleanup at netstack is asynchronous, so a race is possible during the
session restart, where the rule set is yet to be removed when we try
to create a new one for the new session.

Note that this is not common so a retry mechanism of up to 10 times is
likely sufficient.

Fixed: 465160050
Change-Id: I1ca026564d774cb63c00e2bf4e5c97b16a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1442665
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>