[ffx] Allow direct connections to re-materialize the Resolution
Previously, when we set up direct connection in the FHO environment, we
generated a resolution object and then used it for the whole life of the
environment. This is bad because Resolution fixes the IP address of the
device, so if we want to recover from an interruption that changes the
IP, we can't.

The fix is the usual pattern; add a wrapper around Resolution and let it
create a new resolution when needed.

Fixed: 463689273
Test: Manually verified ffx log works when we bounce an emulator and its
      port changes.
Change-Id: I6a6a696433234726e0a40e304459797fb1a71c6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440568
Reviewed-by: Steven Grady <slgrady@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>