[fidl/sockets] Improve FIDL definitions
- Rename "filter" to "matcher" everywhere.
- Update Control.DisconnectIp to match Diagnostics.IterateIp.
- Extract MatcherError and use consistently.
- Remove the `max_sockets` parameter from DisconnectIp.

That last one deserves some justification. While this check is useful in
theory, it requires either allocating a Vec of socket IDs, accepting
that the check is racy, or a relaxation of the check. For now, we're
just going to remove it since it's not a breaking change to add it (or
something similar) back in the future.

Change-Id: I454b3d47753a3b483a57049fec75c8b76a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445948
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>
Fuchsia-Auto-Submit: Peter DeLong <pdelong@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>