[fidl_codec_rust] Add handle abstraction
The FIDL Codec crate is used in a few different contexts:
 - with access to actual (or emulated via overnet) zircon handles
 - with handles provided by the fdomain client library
 - with no actual handles, processing previously captured messages

This has proven to be increasingly tricky. In this CL I've added an
abstraction over handles and fidl-codec-local versions of types involved
in handle use (disposition, info, etc). I've also implemented a FIDL
message header decoder so that we don't have to depend on the FIDL
crate to decode messages.

Change-Id: I8b534c37f6714ebdef4ebb2a399ffbd69a34b45b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1419153
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>