[netstack3] Implement Control.DisconnectIp for TCP
This piggybacks on the existing socket abort functionality, although it
sets a new Aborted error. It is also loosely based on the shutdown API
implementation, but with slightly different semantics since we're
bringing down the whole socket.

For better or worse, this change emulates Linux behavior in terms of
what operations continue to be valid on the SOCK (everything for UDP and
TCP listeners, but nothing for connected TCP sockets). Oddly, this
didn't require any extra work except for validation. In the future, we
may want to revisit this and allow resurrecting more kinds of sockets.

Fixed: 459456549
Change-Id: I3bc16c7980679c49e6d65cd35ec157e06a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445952
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter DeLong <pdelong@google.com>