[network-tun] Use separate dispatcher for netdev
With the recent move away from using Banjo for netdevice the device
adapter started serving NetworkDeviceImpl on a different dispatcher.
Previously it was served by the banjo shims and the dispatchers created
for that purpose.

These dispatchers were special because they were created with a
different owner than default dispatchers. This allowed dispatcher
inlining within the same process, a requirement for the shims to achieve
decent performance. By using these dispatchers and the shim network-tun
also achieved inlining for free.

Create a specific dispatcher used to server NetworkDeviceImpl and set a
different owner for this dispatcher, just like the shim used to do. This
should restore performance to previous or higher levels.

Bug: 466199589
Change-Id: Ife5abaec324d90690ae67d0f412c861653d0772b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446049
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>