[elf_runner] Add allow_exception config for bad_handles policy
This adds a configuration to the elf runner's program block to enable
the bad handle policy with an action of ALLOW_EXCEPTION. This asks the
kernel to raise a non-fatal exception when a violation of the bad handle
policy is detected. Enabling this configuration will allow us to
identify violations of the bad handle policy without changing behavior
(outside of timing) to let us gradually fix issues before ramping up
enforcement of the bad handle policy.

The new syntax for bad handle job policy configuration is:

  program: {
    job_policy_bad_handles: "deny" | "allow_exception",
  }

This also adds docs for this field to the elf runner documentation.

Bug: 42151766

Change-Id: I566bc59357d7190f76965415d388635536deb81a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445735
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>