[assembly][vbmeta] Add footer support
Rather than being standalone, a VBMeta may be found appended to an image
or amongst partition contents. In this case - which will be the case for
pVM VBMeta - it is identified through a "VBMeta footer". We add support
for that here.

Bug: 449201661
Change-Id: I2233806de9f7fb4ed1a8ed3c6ada4b585f7e27db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426935
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>