[wlan][test] Log test start/end in WifiBaseTest
This change adds a logline to the Fuchsia syslog at the setup and
teardown steps for tests that use WifiBaseTest, which helps correlate
syslogs with test logs. The same logging already exists for tests that
inherit FuchsiaBaseTest, so as tests are migrated out of the
//src/testing/end_to_end/antlion dir, they will still retain this
logging.

Fixes: 465799197
Change-Id: I10eb58a5eb8880707eb1498812fe8ee17e0e60f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446687
Commit-Queue: Hayden Nix <haydennix@google.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>