[sestarnix] Add indices_of_set_bits
For most of the current users of ExtensibleBitmap.spans, what is
really wanted is the integers found within the spans representing the
indices of bits set in the ExtensibleBitmap, and a second iteration is
used at each .spans callsite to get these indices. Let's just offer the
iteration yielding the indices natively from ExtensibleBitmap itself.

Fun fact: I drafted this as "set_bit_indices", but then could not pass
self-review with a method with a name that started with "set_" but that
was not mutative. 😐

Change-Id: I9ac21b10673135f9397500effd43ef6030da9c2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447088
Fuchsia-Auto-Submit: Nathaniel Manista <nathaniel@google.com>
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>