[drivers] Add power entities to protocols
Add fields to pass power-related handles in the driver start
args. This will be a way to deliver the handles both to the
driver how, which directly receives the args, and then on to
the driver which gets the args forwarded to it by the driver
host.

Bug: 454169735
Fixed: 462180937

Change-Id: Ie3ff892da28060bbb3a9f0403a8247048ea2f5a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398853
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>