[usb_vsock_host] Remove cid argument from listen()
When I wrote this library I added the ability to listen for incoming
connections either from anywhere, or only from specific source
addresses. I misunderstood POSIX socket semantics and thought the
latter was a standard feature. It is not, and it's a source of some
complexity.

Ripping it out saves us some indirection, and there are no users of the
feature.

Test: unit tests
Change-Id: I6a6a6964c4ed22f68f3442201fa25b4e42959d6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447112
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>