[wlan] Create CoreBaseTestClass and use FuchsiaWlanBaseTest for core end-to-end tests
The largest part of this change is transitioning core end-to-end tests
off BaseTestClass to FuchsiaWlanBaseTest. The latter includes standard
functionality for Fuchsia end-to-end tests so that the classes defined
for WLAN core end-to-end tests don't need to provide that functionality
themselves.

The CoreBaseTestClass makes a connection to
fuchsia.wlan.device.service/DeviceMonitor and asserts exactly one
FuchsiaDevice is available.

Part of this change modifies ConnectTransactionEventHandler to take an
asyncio.AbstractEventLoop as an argument to step around the honeydew
async limitations. This will be unwound in a later change.

Also, a basic ping function is added ConnectionBaseTestClass for
FirmwarePowerModesTest. As other tests need this method, we may need to
move it elsewhere, even into honeydew as an affordance.

Change-Id: I11624bc691d1741928c6e2c3e6cad097230c06c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441530
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>