[security] Update Scrutiny structured config tests
This introduces a component_tree_config file to allow structured config
tests to use dynamic routing information. This is needed because the
structured_config tests create a test product bundle, which seem to
inherit a number of changes to core and is not well-insulated.

Pointing to the platform-level config might work, but this change will
opt to use a test-specific copy.

Bug: b/468009955
Change-Id: I01fa03a2a0fe14bc863a314e0b0cec22cda6dabf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446192
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Eric Stone <ecstone@google.com>