[fuchsia-controller] Cancel pending tasks upon external cancellation
The FidlClient creates background tasks to wait on a handle being ready
to read, a staged message becoming available, or an epitaph on the
handle. Without structured concurrency, cancellation of the task that
created those tasks would result in the background tasks printing errors
that they were still pending when destroyed.

This change uses asyncio.TaskGroup to create the background tasks in a
manner that propagates the cancellation of the parent task to each
background task.

Change-Id: Icdc21f494ab80f9e518b730d2a2ca196baef79c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446968
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>