[security] Add platform component_tree_config
This file will allow platform-defined components which use dynamic
routing features to tell Scrutiny how to incorporate their capabilities
into Scrutiny's model. In this current state, it enforces that code
which dynamically handles capabilities be able to represent the set of
capabilities in this static config file.

For example, this is needed for config capabilities presented from a
dynamic dictionary to successfully route for the structured config
model.

To soft-migrate OOT usages, this change will add a new field to
represent a list of component tree configs without removing the existing
single-file field.

Bug: b/468009955
Change-Id: I8b3c23eef37b63854c8ede61e785ab389364c967
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446188
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>