[starnix] Creates a builder-based API for dynamic_thread_spawner.
This only modifies `dynamic_thread_spawner`, the rest of the
refactor will be done piecemeal in separate changes.

In anticipation of more options to the zoo of "spawn" methods on
`DynamicThreadSpawner`, I refactored its API to use a typestate-based
builder pattern. The typestate builder only ever allows valid states to
be represented when building a spawn request. This makes using it very
straightforward. The tradeoff is a somewhat verbose declaration of the
builder methods.

This required a lot of noisy template parameterization, which is
something that should be avoided unless absolutely necessary. On the
flip side, it makes it impossible for the user of this API to
misconfigure a spawn request.

I experimented with a builder that does not use typestate. It turned out
that the API has enough confusing corner cases that it would be
decidedly more difficult to use correctly.

For this reason I am now convinced that typestate, however noisy, is the
correct choice between the alternatives.

Another good outcome of this refactor is that the DynamicThreadSpawner
API was split into two APIs with narrower concerns: the Builder API
deals only with configuring the spawn request correctly. The Spawner API
only deals with spawning tasks.

Bug: 465144050
Change-Id: Ie57593a344504f869bcef2efc7905ebb7f9d17d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1439505
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>