[kernel][arm] Use x28 instead of x29 when entering user-mode.
A recent change, here:

https://fuchsia-review.git.corp.google.com/c/fuchsia/+/1446950

altered the way that we load the initial registers when entering user
mode for the first time on a thread.  It attempted to better align the
way the registers were loaded with the C definition of the iframe
structure in order to allow the structure's layout to be changed (with
some restrictions) without invalidating the assembly code which
depended on its layout.

In the process, a small optimization was made to avoid a single
register load, and to use a load-pair instruction instead.  This
involved moving pointer to the iframe structure held in x0 into x29
for the general register load sequence.

A downside to this approach stems from the fact that that x29 is the
frame pointer.  If we overwrite the current frame pointer with the
iframe pointer, and then we take an exception while restoring
registers, any attempt to produce a backtrace is going to be severely
compromised.  Prior to the re-organization, x29 was allowed to hold
the current frame pointer all of the way until the final load,
followed by eret, minimizing the chance of this.

So, switch from using x29 as our iframe pointer to using x28 instead.
x29 is now allowed to keep the current valid frame pointer value right
up until the final load-pair followed by the eret, minimizing the
window for failure as was done before.

Change-Id: I6f7efca23c551e31e57ac5b246861f16f090c5eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1448072
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>