[components] DirConnector in fuchsia.component.sandbox to match internal
fuchsia.component.sandbox was created before much of the underlying
implementation had been migrated to the new runtime capabilities. During
this migration we discovered that some use cases called for being able
to pass flags and a subdir when opening a dir connector, despite the
preexisting API not supporting this.

A replacement for fuchsia.component.sandbox is in the works that will
address this along with other shortcomings, but some users are blocked
on the ability to use the full directory connector API.

To unblock these users in the short term, this commit reworks the
DirConnector APIs in fuchsia.component.sandbox to support passing these
arguments when opening the directory connector.

Change-Id: I765ab5b4c85319d7c0aa7cabf97a341faf9dd023
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1390873
API-Review: Hunter Freyer <hjfreyer@google.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>