[sshd-host]: better handle logging of sshd exit code
When the service OnStop method is called, check the exit code of the
sshd child component, if it is 255 that's the signal that the client
terminated the connection.

Test: Invoke ffx log --component klog --component core/sshd-host. Then
invoke in another terminal `ffx -c ssh.controlmaster='none' log` and
wait until logs appear. Then CtrlC the log and note that the error is
not present at the WARN level for sshd-host.
Change-Id: Iafdfd054218b73796e2bfc0ac75a16516a6a6964

Change-Id: Iafdfd054218b73796e2bfc0ac75a16516a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1449298
Reviewed-by: Steven Grady <slgrady@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>