[starnix] SysRq crashes Starnix rather than rebooting
When "c" is written to SysRq, panic the Starnix kernel directly
rather than rebooting the whole device. This is useful in
testing a recovery / reporting logic and avoiding the confusion
between the different failure modes.

The existing rebooting logic is moved to "b" which seems to be a better
fit.

Test: starnix_sysrq_integration_test

Bug: 394953531
Change-Id: I11705961a7b90a90ac3ce713de93b5c88a5ea2f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447888
Commit-Queue: Maria Glukhova <mariagl@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>