[rust][zx] Update tests to satisfy bad handle policy
This removes tests that interact with system calls only to the extent of
triggering the bad handle policy and configures the suite's bad handle
policy to raise exceptions.

If we want to exercise the syscall bindings for wrappers that interact
with privileged or otherwise specialized object types we could use the
fake-object library to interpose on the zx_.. symbols.

Bug: 42151766

Change-Id: Id4e91fb78ee5a4836cdd8b43197cd321b6ce9c52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446196
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: James Robinson <jamesr@google.com>