[component_manager] Reduce number of items in realm_query::manifest_test to speed it up
This adds assertions on the size of the serialized manifest, and then
reduces the number of items it contains so that it fits in more than
one, but not more than 3, zircon channel messages.

This should significantly speed up this test on some platforms.

Change-Id: Ida800c9cd6025da542d639d93a061074e7f88b0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1449016
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>