[wlan] Use legacy chanspec as fallback in connect/roam.
If a connect or roam attempt encounters a WLAN channel that fails its
internal channel to chanspec conversion, attempt to fallback on the
legacy chanspec conversion code. In one specific case (channel 165 with
wide bandwidths), the driver and firmware has been seen to fail connect
when attempting to configure for the wider bandwidth; if we encounter
other cases like this, the driver can attempt to use the old chanspec
logic rather than immediately failing.

Bug: 462439885
Change-Id: Ie85dbd562838c18e46caef609a616f2679230f4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1446010
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Karl Ward <karlward@google.com>