[arch][arm] tweak global cache coherency context manager to add additional barriers
In the case of the cpu signalling that it doesn't need the data cache to
be written out before flushing the i cache, make sure a data barrier is
always present before potentially flushing the i cache to make sure it
at least gets the freshest of data.

In the case where the cpu signals that it does not need an icache
invalidation to synchronize the caches, make sure an ISB is issued.

Change-Id: I6917b7307689ddb1390f4d9e40ee553935612c4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447110
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>