[userboot][core-tests] Expose when midr.txt is emitted
Currently we only emit midr.txt when ZX_DEBUG_ASSERT_IMPLEMENTED is
nonzero. However, we are trying to run core tests in a configuration
where it is 0 (*-no_kernel_debug) and are failing to skip the
ProcessTest.GetInfoRuntime* tests in that context since there is no
midr.txt to make the "is Cavium?" check.

Instead of having the core test logic hardcode an expectation around
when a midr.txt is present, we introduce a constant in libuserabi that
can be consulted instead. When it indicates that no midr.txt is present
then we will always skip this test.

Fixed: 467750183
Change-Id: I46ef44decd787450b812b686fb97a0e38a10f2ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447011
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>